pyproject.toml
[build-system]
requires = ["hatchling>=1.18"]
build-backend = "hatchling.build"

[project]
name = "aws-cli-mcp"
version = "0.1.0"
description = "Model Context Protocol server that exposes vetted AWS CLI and SDK tools"
authors = [
  { name = "Shuto", email = "" }
]
requires-python = ">=3.11"
license = { text = "MIT" }
readme = "README.md"
dependencies = [
  "boto3>=1.34",
  "fastmcp>=2.11.4",
  "pydantic>=2.6",
  "python-dotenv>=1.0",
  "jsonschema>=4.22",
  "pyyaml>=6.0",
  "httpx>=0.27",
  "PyJWT[crypto]>=2.8",
  "starlette>=0.37",
  "uvicorn>=0.29",
]

[project.optional-dependencies]
dev = [
  "pytest",
  "pytest-cov",
  "ruff",
  "mypy",
  "pip-audit>=2.7",
]

[tool.hatch.build.targets.wheel]
package-dir = "src"
packages = ["aws_cli_mcp"]

[tool.hatch.build.targets.sdist]
include = [
  "src/aws_cli_mcp",
  "README.md",
  "docs",
  "pyproject.toml",
  ".env.example",
]

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.ruff.lint.per-file-ignores]
"tests/*.py" = ["E501"]

[tool.pytest.ini_options]
pythonpath = ["src"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
asyncio_default_test_loop_scope = "function"

[tool.mypy]
python_version = "3.11"
show_error_codes = true
warn_unused_ignores = true

[[tool.mypy.overrides]]
module = [
  "boto3",
  "boto3.*",
  "botocore",
  "botocore.*",
  "jsonschema",
  "jsonschema.*",
  "yaml",
  "yaml.*",
]
ignore_missing_imports = true
